Move config message factory
Move the config message factory from production to test code; it is not
production code, so it does not belong there.

Additionally, correct the docker compose file for the config tests, such
that the source directory is correct.
